Add endpoint to fetch all blogs that belong to related user

+++ backend/src/main/java/com/inklet/blog/blog_backend/entity/Blog.java
@@ -13,7 +13,7 @@ public class Blog {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "title")
     private String title;
@@ -58,17 +58,17 @@ public class Blog {
     }
 
     //For updating a Blog
-    public Blog(int id, String title, String content) {
+    public Blog(Long id, String title, String content) {
         this.id = id;
         this.title = title;
         this.content = content;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
